chore(refactore): #1160 moved profile, features, help, privacy components under main module
