+++ web/src/app/main/components/legal/privacy/privacy.component.ts
