+++ web/src/app/main/components/legal/terms-conditions/terms-conditions.component.ts
