+++ web/src/app/main/main-routing.module.ts
@@ -3,18 +3,18 @@ import { Routes, RouterModule } from '@angular/router';
 
 // Dashboard hub components
 import { MainComponent } from './main.component';
-import { HomepageComponent } from '../homepage/homepage.component';
+import { HomepageComponent } from './components/homepage/homepage.component';
 import { CreateProjectComponent } from '../projects/create/create.component';
 import { ViewProjectComponent } from '../projects/view/view.component';
 import { ViewProjectResolver } from '../core/resolvers/view-project.resolver';
 import { EditProjectComponent } from '../projects/edit/edit.component';
 import { EditProjectResolver } from '../core/resolvers/edit-project.resolver';
 import { PrivateProjectsComponent } from '../projects/private/private.component';
-import { ProfileComponent } from '../profile/profile.component';
-import { FeaturesComponent } from '../features/features.component';
-import { HelpComponent } from '../help/help.component';
-import { TermsConditionsComponent } from '../legal/terms-conditions/terms-conditions.component';
-import { PrivacyComponent } from '../legal/privacy/privacy.component';
+import { ProfileComponent } from './components/profile/profile.component';
+import { FeaturesComponent } from './components/features/features.component';
+import { HelpComponent } from './components/help/help.component';
+import { TermsConditionsComponent } from './components/legal/terms-conditions/terms-conditions.component';
+import { PrivacyComponent } from './components/legal/privacy/privacy.component';
 
 import { AuthGuard } from '../core/guards/authentication.guard';
